backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
cryptography==42.0.8
redis>=5.0.0
pydantic-settings==2.3.4
python-multipart==0.0.9
aiosmtplib==3.0.1
email-validator==2.1.1
httpx==0.27.0
python-dotenv==1.0.1
bcrypt==4.1.3
qrcode==8.2
